+++ src/main/php/lang/reflect/ClassParser.class.php
@@ -31,7 +31,8 @@ class ClassParser extends \lang\Object {
     if ('self' === $type) {
       return XPClass::forName($context);
     } else if ('parent' === $type) {
-      return XPClass::forName($context)->getParentclass();
+      if ($parent= XPClass::forName($context)->getParentclass()) return $parent;
+      throw new IllegalStateException('Class does not have a parent');
     } else if (false !== strpos($type, '.')) {
       return XPClass::forName($type);
     } else if (isset($imports[$type])) {